Extract ForwardingRecordCursor
