+++ presto-hive/src/test/java/io/prestosql/plugin/hive/util/TestForwardingRecordCursor.java
@@ -1,0 +1,28 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.prestosql.plugin.hive.util;
+
+import io.prestosql.spi.connector.RecordCursor;
+import org.testng.annotations.Test;
+
+import static io.prestosql.spi.testing.InterfaceTestUtils.assertAllMethodsOverridden;
+
+public class TestForwardingRecordCursor
+{
+    @Test
+    public void testAllMethodsOverridden()
+    {
+        assertAllMethodsOverridden(RecordCursor.class, ForwardingRecordCursor.class);
+    }
+}
